Updates
- cleanup hamiltonian cpp/hpp files
- change order of arguments for mdarray constructor to be more consistent
