+++ src/Hamiltonian/local_operator.hpp
@@ -27,13 +27,15 @@
 
 #include "../SDDK/GPU/acc.hpp"
 #include "../SDDK/memory.hpp"
-#include "../smooth_periodic_function.hpp"
 #include "spfft/spfft.hpp"
+#include "typedefs.hpp"
 
 /* forward declarations */
 namespace sirius {
 class Potential;
 class Simulation_context;
+template <typename T>
+class Smooth_periodic_function;
 }
 namespace sddk {
 class FFT3D;
@@ -87,13 +89,13 @@ class Local_operator
     sddk::mdarray<double, 1> pw_ekin_;
 
     /// Effective potential components on a coarse FFT grid.
-    std::array<Smooth_periodic_function<double>, 4> veff_vec_;
+    std::vector<Smooth_periodic_function<double>> veff_vec_;
 
     /// Temporary array to store [V*phi](G)
     sddk::mdarray<double_complex, 1> vphi_;
 
     /// LAPW unit step function on a coarse FFT grid.
-    Smooth_periodic_function<double> theta_;
+    std::unique_ptr<Smooth_periodic_function<double>> theta_;
 
     /// Temporary array to store psi_{up}(r).
     /** The size of the array is equal to the size of FFT buffer. */
