+++ src/sirius.h
@@ -35,6 +35,8 @@
 using json = nlohmann::json;
 
 #include "input.hpp"
+#include "simulation_context.hpp"
+#include "Hamiltonian/local_operator.hpp"
 #include "radial_solver.hpp"
 #include "SHT/sht.hpp"
 #include "SHT/gaunt.hpp"
